Move consolidation to summarycode
