+++ tests/summarycode/data/plugin_consolidate/component-package-build-expected.json
