+++ tests/summarycode/data/plugin_consolidate/component-package-build/component/src3
