+++ tests/summarycode/data/plugin_consolidate/component-package-build/package/src2
