+++ tests/summarycode/data/plugin_consolidate/component-package-expected.json
