+++ tests/summarycode/data/plugin_consolidate/component-package/component/src2
