+++ tests/summarycode/data/plugin_consolidate/component-package/package/src2
