+++ tests/summarycode/data/plugin_consolidate/e2fsprogs.json
