+++ tests/summarycode/data/plugin_consolidate/license-holder-rollup-expected.json
