+++ tests/summarycode/data/plugin_consolidate/license-holder-rollup/clear-majority/dir1/file1
