+++ tests/summarycode/data/plugin_consolidate/license-holder-rollup/no-majority/foo/baz
