+++ tests/summarycode/data/plugin_consolidate/multiple-same-holder-and-license-expected.json
