+++ tests/summarycode/data/plugin_consolidate/nested-packages/package.json
