+++ tests/summarycode/data/plugin_consolidate/package-files-not-counted-in-license-holders/package/src3
