+++ tests/summarycode/data/plugin_consolidate/package-fileset-expected.json
