+++ tests/summarycode/data/plugin_consolidate/package-manifest-expected.json
