+++ tests/summarycode/data/plugin_consolidate/package-manifest/package.json
