+++ tests/summarycode/data/plugin_consolidate/package/package.json
