+++ tests/summarycode/data/plugin_consolidate/package/src1
