+++ tests/summarycode/data/plugin_consolidate/report-subdirectory-with-minority-origin-expected.json
