+++ tests/summarycode/data/plugin_consolidate/report-subdirectory-with-minority-origin/minority_holder/a
