+++ tests/summarycode/data/plugin_consolidate/return-nested-local-majority/foo/bax
