+++ tests/summarycode/data/plugin_consolidate/return-nested-local-majority/foo/local-majority/baz
