+++ tests/summarycode/data/plugin_consolidate/zlib-expected.json
