+++ tests/summarycode/data/plugin_consolidate/zlib.json
