+++ tests/summarycode/test_plugin_consolidate.py
